Fix Horizen transaction output checks
This commit fixes the support for Horizen blockchain.

The previous implementation was only taking into account outputs including the replay protection.

That made sense as replay protection is mandatory, but it didn't take into account the release of Zendoo and the fact that sidechain certificates include Backward Transfers (outputs) without replay protection.